added float64 conversion for pressure and temperature data

+++ src/polly2scc.py
@@ -1267,8 +1267,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol355 = np.nanmean(
                     beta_pi_rayleigh(
                         355,
-                        pressure=pData['pressure'][refMask355],
-                        temperature=pData['temperature'][refMask355] + 273.16))
+                        pressure=np.float64(pData['pressure'][refMask355]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask355] + 273.16)))
                 refBscRatio355 = refVal_355 / refBscMol355 + 1
 
                 # 0: monte_carlo;
@@ -1326,8 +1327,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol355 = np.nanmean(
                     beta_pi_rayleigh(
                         355,
-                        pressure=pData['pressure'][refMask355],
-                        temperature=pData['temperature'][refMask355] + 273.16))
+                        pressure=np.float64(pData['pressure'][refMask355]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask355] + 273.16)))
                 refBscRatio355 = refVal_355 / refBscMol355 + 1
 
                 # 0: monte_carlo;
@@ -1394,8 +1396,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol532 = np.nanmean(
                     beta_pi_rayleigh(
                         532,
-                        pressure=pData['pressure'][refMask532],
-                        temperature=pData['temperature'][refMask532] + 273.16))
+                        pressure=np.float64(pData['pressure'][refMask532]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask532] + 273.16)))
                 refBscRatio532 = refVal_532 / refBscMol532 + 1
 
                 # 0: monte_carlo;
@@ -1453,8 +1456,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol532 = np.nanmean(
                     beta_pi_rayleigh(
                         532,
-                        pressure=pData['pressure'][refMask532],
-                        temperature=pData['temperature'][refMask532] + 273.16))
+                        pressure=np.float64(pData['pressure'][refMask532]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask532] + 273.16)))
                 refBscRatio532 = refVal_532 / refBscMol532 + 1
 
                 # 0: monte_carlo;
@@ -1521,9 +1525,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol1064 = np.nanmean(
                     beta_pi_rayleigh(
                         1064,
-                        pressure=pData['pressure'][refMask1064],
-                        temperature=pData['temperature'][refMask1064] +
-                        273.16))
+                        pressure=np.float64(pData['pressure'][refMask1064]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask1064] + 273.16)))
                 refBscRatio1064 = refVal_1064 / refBscMol1064 + 1
 
                 # 0: monte_carlo;
@@ -1581,9 +1585,9 @@ class polly_2_earlinet_convertor(object):
                 refBscMol1064 = np.nanmean(
                     beta_pi_rayleigh(
                         1064,
-                        pressure=pData['pressure'][refMask1064],
-                        temperature=pData['temperature'][refMask1064] +
-                        273.16))
+                        pressure=np.float64(pData['pressure'][refMask1064]),
+                        temperature=np.float64(
+                            pData['temperature'][refMask1064] + 273.16)))
                 refBscRatio1064 = refVal_1064 / refBscMol1064 + 1
 
                 # 0: monte_carlo;
